+++ main-1-2.cpp
@@ -2,7 +2,7 @@
 
 int main() {
 
-extern double array_mean(double array[], int n);
+extern double array_mean(int array[], int n);
 
 double mainArray[] = {};
 int arraySize = sizeof(mainArray)/sizeof(mainArray[0]);
